Extract rollup warning handler to a separate file
